[FIX] mail: empty push notification when tracking record changes
Before this commit:

Push notifications were empty when a user tracked record changes (e.g., stage
changes in a task). This happened because trackingValues were not included
in the notification message.

After this commit:

trackingValues are now correctly added to the notification message, ensuring
that push notifications display the relevant record change details.

Task-4510508
